rename: 페이지명 notification -> alerts로 수정
